NecrosisApi/requirements.txt
Django~=5.1.2
djangorestframework~=3.15.2
ultralytics~=8.3.25
torch~=2.5.1
opencv-python~=4.10.0.84
numpy~=2.1.2
pillow~=11.0.0